drivers: serial: xuartps: fixing set_baudrate function
The Xilinx xuartps driver fails to find the correct pre-scaler
parameters for the BAUDDIV and BAUDGEN registers. This fix
corrects that behaviour. The reason why it fails is due to
the stop condition for the search being ill-conditioned.
The stop condition is based on a baudrate error, and this
was chosen to be 3 percent. This would cause the parameter
search to stop too early returning ill-condotioned pre-scaler
parameters. This has been corrected by setting the stop condition
to be 3 permille instead of 3 percent. This change was tested
with an AMD Zynq Ultrascale+ MPSoC ZU3EG.

(cherry picked from commit 5599bf59148c4e49327dbeab03070f5e9c96edbb)

Original-Signed-off-by: Kent Stark Olsen <kent.stark.olsen@gmail.com>
GitOrigin-RevId: 5599bf59148c4e49327dbeab03070f5e9c96edbb
Cr-Build-Id: 8719994525146275441
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8719994525146275441
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I8ff25d2e223631939728bdc8346d6c6045eac40a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6371308
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>