tests: Bluetooth: Tester: Refactor ACSC events for BAP unicast
This commit refactors when the BTP_ASCS_EV_OPERATION_COMPLETED
is sent to the upper tester. Previously it was send
in the stream changed callbacks, but that would send
the events to the BAP Unicast Server when it did not initiate
any changes, and would not properly match when the control
point operation was completed (which happens before the
ASE state changes).

This change makes it easier to determine when an operation was
actually completed, and easier to separate that from a
ASE state change which is another type of event.

(cherry picked from commit ae4c6fe6097033fde56e2e1c0a3637f0985ea772)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: ae4c6fe6097033fde56e2e1c0a3637f0985ea772
Cr-Build-Id: 8715532624604821329
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8715532624604821329
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I78f2357a737680b0ada16f2e63a781042ef95f43
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6519454
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>