llext: stricter alignment checks for MPU/MMU regions
This patch moves the alignment checks for MPU and MMU regions to the
beginning of the llext_copy_region() function. This is done to ensure
that the correct region alignment and size are used even in the case
where the region is reused from the ELF file buffer, avoiding MMU/MPU
configuration issues.

This also relaxes the same checks for regions that are accessed by the
kernel only (e.g. symbol and string tables), which do not need special
MMU/MPU treatment. This exposed an inconsistency in the MMU code, which
was setting the permission on the correct regions, but later restoring
the default permissions on every region, including the now-unaligned
ones.

(cherry picked from commit e5594891ac156ae9419f1017bc78d9d17c09d366)

Original-Signed-off-by: Luca Burelli <l.burelli@arduino.cc>
GitOrigin-RevId: e5594891ac156ae9419f1017bc78d9d17c09d366
Cr-Build-Id: 8719745383977653153
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8719745383977653153
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I19a64157c556a36023b1d1028bb7d7d4cad23dd9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6384746
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>