net: lib: tls_credentials: Rename TLS_CREDENTIAL_SERVER_CERTIFICATE
TLS_CREDENTIAL_SERVER_CERTIFICATE credential type is misleading, as in
fact it just represents a public certificate, it does not matter if the
certificate belongs to a server or a client. And actually, it was
already used in-tree for clients as well, for example in LwM2M.

Therefore rename the credential type to a more generic
TLS_CREDENTIAL_PUBLIC_CERTIFICATE and deprecate the old one.

(cherry picked from commit a61287e8766fe14d0055fba796c56948c483eaaf)

Original-Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>
GitOrigin-RevId: a61287e8766fe14d0055fba796c56948c483eaaf
Cr-Build-Id: 8719768033269589329
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8719768033269589329
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I830f4a7b3fe401a8b8cb27c468ab24ba733a8ff1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6382730
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>