pm: device_runtime: Allow to use a dedicated wq
Device runtime is using the system workqueue to do operations
that are mostly blockers (suspend a device). This should not happen.

This commit adds an option to use dedicated queue for the device runtime
async operations.

The test for this API was assuming that the system workqueue
priority (which is cooperative) so we need to change the test priority
to be lower than the device runtime workqueue priority.

(cherry picked from commit 64e38bab556c6a6a4810109258b825d0d9a2c288)

Original-Signed-off-by: Flavio Ceolin <flavio@hubblenetwork.com>
GitOrigin-RevId: 64e38bab556c6a6a4810109258b825d0d9a2c288
Cr-Build-Id: 8713584789952251569
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8713584789952251569
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Icc459b636422c220ef4af74f727cf5d6c457fc37
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6599803
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>