json: improve parsing and serializing of 'float' and 'double'
Up to now, the handling of type float was offloaded to the users of the
JSON utility, with the token JSON_TOK_FLOAT.

Improve handling of floating point types and support the types 'float'
and 'double' in a built-in way so that they can be directly parsed to
and serialized from variables (of type float or double).

The types are serialized in the shortest representation, either as a
decimal number or in scientific notation:
  * float (with JSON_TOK_FLOAT_FP): encoded with maximal 9 digits
  * double (with JSON_TOK_DOUBLE_FP): encoded with maximal 16 digits
  * NaN, Infinity, -Infinity: encoded and decoded as:
    {"nan_val":NaN,"inf_pos":Infinity,"inf_neg":-Infinity}

Enable the floating point functionality with the Kconfig option:
  JSON_LIBRARY_FP_SUPPORT=y

It requires a libc implementation with support for floating point

(cherry picked from commit fc37c02eb134e19541a2559a74fbaf4b339fa2d0)

Original-functions: strtof(), strtod(), isnan() and isinf().
Original-Fixes: #59412
Original-Signed-off-by: Christoph Winklhofer <cj.winklhofer@gmail.com>
GitOrigin-RevId: fc37c02eb134e19541a2559a74fbaf4b339fa2d0
Cr-Build-Id: 8718024041812476529
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8718024041812476529
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I7762fb05088e003655b90744ec5262633dc32d2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6444617
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>