json: Fix calculation of object size
The calculation of the object size may be incorrect when the size of
a field is smaller than the struct alignment. When such a struct is
used in an array field, the decoded object contains wrong values.

The alignment influences the object size. For example the following
struct has a calculated object size of 8 bytes, however due to
alignment the real size of the struct is 12 bytes:

struct test_bool {
  bool b1; /* offset 0, size 1 */
           /* 3-byte padding */
  int i1;  /* offset 4, size 4 */
  bool b2; /* offset 8, size 1 */
           /* 3-byte padding */
};

This commit changes the object size calculation and computes the size
with the offset and size of the last field in the struct (rounded up
by the struct alignment).

(cherry picked from commit 061e34597245e87f9dd218cae1fad261f16782dd)

Original-Fixes: #85121
Original-Signed-off-by: Christoph Winklhofer <cj.winklhofer@gmail.com>
GitOrigin-RevId: 061e34597245e87f9dd218cae1fad261f16782dd
Cr-Build-Id: 8718159936872580593
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8718159936872580593
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I0841deb04c6168a405a86a9e9b04a011ccb1f409
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6441514
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>