arch/x86: Verify shadow stack pointers
Currently, it's permitted to have threads that don't have a shadow
stack. When those are run, shadow stack is disabled on the CPU. To
identify those, the thread `shstk_addr` member is checked.

This patch adds an optional check, behind
CONFIG_X86_CET_VERIFY_KERNEL_SHADOW_STACK, that checks if an outgoing
thread has this pointer NULL with shadow stack currently enabled on
the CPU, meaning a 1) bug or 2) some attempt to tamper with the pointer.

If the check fails, k_panic() is called. Note that this verification is
not enough to guarantee `shstk_addr` can't be tampered with. For
instance, it only works on a running thread. Ideally, all threads should
be shadow stack capable, so a missing `shstk_addr` would simply be a
hard fault, but that is still to come.

(cherry picked from commit 86b401075ac8b89a6cee609d7f2414484041e692)

Original-Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>
GitOrigin-RevId: 86b401075ac8b89a6cee609d7f2414484041e692
Cr-Build-Id: 8704751585743389537
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8704751585743389537
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I5281c282d4e812f9a7a1b3af4bf6b6d3c818e2c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6907918
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>