arch/x86: Support shadow stack on IRQ
For IRQs, shadow stack support a mechanism similar to the Interrupt
Stack Table (IST) for x86_64: a table, indexed by the IST index, pointing
to a 64 byte table in memory containing the address of seven shadow stacks
to be used by the interrupt service routines.

This patch adds support to this mechanism. It is worth noting that, as
Zephyr may exit from an interrupt by going to a different thread than
the one that was interrupted, some housekeeping is done to ensure that
the necessary shadow stack tokens are on the right shadow stack before
return from the interrupt.

(cherry picked from commit 246cd228f2921ab72e5912c9ae75f964fe1f0759)

Original-Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>
GitOrigin-RevId: 246cd228f2921ab72e5912c9ae75f964fe1f0759
Cr-Build-Id: 8704751585743389537
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8704751585743389537
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I4fc60190df01f389a50ec7532346a68e500003dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6907915
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>