Bluetooth: Classic: L2CAP: Discard the response with a invalid `ident`
In current implementation, the `ident` of the command response is not
checked whether it is same as the `ident` sent by the L2CAP command.

Keep the `ident` of the command. If the received `ident` of the
command response is not aligned with the `ident` of the command,
discard the response silently. Or, clear the saved `ident`.

(cherry picked from commit 384e31977f594522330651436c71f95391d3674f)

Original-Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>
GitOrigin-RevId: 384e31977f594522330651436c71f95391d3674f
Cr-Build-Id: 8718748814701089329
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8718748814701089329
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I6466470af13e0319e431c6ac0d72fb320f98c832
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6423572
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>