cmake: yaml: switch to yaml for intermediate file
The intermediate file used to expand the generator expressions is now in
YAML format. This allows for a more robust handling of the data, as the
single quoted strings are a lot easier to escape: inside them every
character is a literal except for two single quotes (which intuitively
map to a literal single quote).

When saving a simple YAML file without genexes, the single quotes in
strings need to be escaped. However, doing so when saving the
intermediate file would make it harder to properly escape them later,
since it could be possible that the expansion of the generator
expressions would introduce new single quotes. For this reason, when
genexes are enabled, the escaping is now done in a single pass inside
the yaml-filter.cmake script.

(cherry picked from commit 90aa937e3a036150ebb1eb3b01a7eec658e518fa)

Original-Signed-off-by: Luca Burelli <l.burelli@arduino.cc>
GitOrigin-RevId: 90aa937e3a036150ebb1eb3b01a7eec658e518fa
Cr-Build-Id: 8717978743039779921
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8717978743039779921
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I274544e45357b4355656964e6901841fcbb23c43
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6446322
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>