arch: common: Make nocache region loadable
The `nocache` is not loadable, thus data stored therein cannot be
initialized by the startup code. This might be needed in special
cases. E.g. One might have a buffer which one wants to DMA into,
and which is a member of a struct. Other members of the struct one
may want to have initialized by the startup code.
The buffer thus should be placed in the `nocache` region, but for
the other members of the buffer to be initialized by the startup
code, the `nocache` region needs to be loadable.

Fix it by making the `nocache` region loadable. Adding a KConfig
symbol to do this optionally was considered, but deemed unnecessary
during the PR.

(cherry picked from commit f9168ae46409bb4a7908c48f4385cb3ac4bc6888)

Original-Signed-off-by: Julian Achatzi <mail@achatzi.pro>
GitOrigin-RevId: f9168ae46409bb4a7908c48f4385cb3ac4bc6888
Cr-Build-Id: 8716982175448322353
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8716982175448322353
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I91a4880e9d2eaf79576a3e4ec567a0c3341d625f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6474698
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>