arch: arm: fix ci failure
This is to fix ci failure caused by commit d5d4d57

In the above commit, it's wrongly assumed that ROM_ADDR and RAM_ADDR
are same **only** when CONFIG_XIP is not defined.
Since aarch64 was taken as reference in the review,
we missed to see that CONFIG_FLASH_SIZE is also part of the condition
to define ROM_ADDR as RAM_ADDR.
CONFIG_XIP is not defined for qemu_cortex_a9/xc7z007s however,
CONFIG_FLASH_SIZE is not 0 so the assumption fails
and the required reset to RAM_ADDR is not applied leading to the test
failure.

(cherry picked from commit d4d09225cabf923c3bbc5151c10a88912dd404eb)

Original-Signed-off-by: Sudan Landge <sudan.landge@arm.com>
GitOrigin-RevId: d4d09225cabf923c3bbc5151c10a88912dd404eb
Cr-Build-Id: 8718771466577921585
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8718771466577921585
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ife9a7cb1b698145cda009a3bfc1a130c88c68a3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6421600
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>