doc: _extensions: zephyr: Use binding title in supported hw table
Use the title property from bindings in the supported hardware table
if available.

Create <abbr> elements for acronyms and abbreviations in titles.
Abbreviations are not created when the description field is used as
a title, since existing text in bindings was not created with this
in mind, and parentheses are used in the text in ways that are
incompatible with the auto-abbreviating logic.

(cherry picked from commit b2451ca10931afad05cf71da26f186db2f002b2d)

Original-Signed-off-by: Aksel Skauge Mellbye <aksel.mellbye@silabs.com>
GitOrigin-RevId: b2451ca10931afad05cf71da26f186db2f002b2d
Cr-Build-Id: 8715940311004562449
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8715940311004562449
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I5fcce9f7afade9a6c4a841ea21a2bac2cf2dfd28
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6509568
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>