Bluetooth: Controller: Fix nRF CCM disable on connection event abort
Fix missing nRF CCM disable on connection event abort.

There can be a problem on nRF SoC for example when a S8
"encrypted" reception is aborted, and a 2M "cleartext"
reception starts; slow CCM (that is not stopped as part of
radio disable) will corrupt a fast received "cleartext" when
the same current free rx buffer is reused in the Controller.
This is not a problem when the connection being abort-ee is
on a faster PHY than the abort-er.

(cherry picked from commit 2fc54ae3bb74cdb2d2bc81d9248a809276cd33e4)

Original-Signed-off-by: Vinayak Kariappa Chettimada <vich@nordicsemi.no>
GitOrigin-RevId: 2fc54ae3bb74cdb2d2bc81d9248a809276cd33e4
Cr-Build-Id: 8714694602652640193
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8714694602652640193
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I588f639ff7b423466f2eebb1c1dac08860555d21
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6558085
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>