posix: timers: correct pointer passed to k_mem_slab_free()
If it is not possible to create a timer in timer_create(),
then the timer_obj associated with the timer must be freed.

However, the address of the pointer was mistakenly being
passed to k_mem_slab_free() rather than simply the
the pointer.

This caused a crash in tests which can easily be avoided
by passing the pointer rather than the address of the
pointer.

(cherry picked from commit 82d564a3547017e0f0aadcdd4a16a03c784cf635)

Original-Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>
GitOrigin-RevId: 82d564a3547017e0f0aadcdd4a16a03c784cf635
Cr-Build-Id: 8717435159221032833
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8717435159221032833
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I1beb44f3456a5052c1c91e567c16bda5bf237d8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6462163
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>