arch: arm: mpu: Fix alignment check for iccarm
The IAR C/C++ compiler can't resolve a cast on a pointer
to a constant expression. This is used in the
_ARCH_MEM_PARTITION_ALIGN_CHECK macro.

This fix checks if an IAR compiler is used and disables the
start-alignment check if it is.

This was already fixed for v7 in rev d34d554d.

(cherry picked from commit 8cfd84986d93b6d929f517e7758c974c5a9623b4)

Original-Signed-off-by: Lars-Ove Karlsson <lars-ove.karlsson@iar.com>
GitOrigin-RevId: 8cfd84986d93b6d929f517e7758c974c5a9623b4
Cr-Build-Id: 8713630084945030369
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8713630084945030369
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I1f3a31486fa251bb071997bfe8b80729835a3113
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6583403
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>