Bluetooth: BAP: Sink: Move mod_src_param to RAM
The mod_src_param was several places stored on the stack.
However this is a complex paramater struct that has
2 Kconfig options that can significantly increase the size,
and the maximum size of the parameter is nearly 8 KiB, and
was always place the in the BT RX thread's stack.

For this reason, the param is now stored in a single
static variable in RAM instead, so that the BT RX thread's stack
does not need to be increased based on the Kconfig options,
as that is quite difficult for users to be aware of.

The add_src_param has been left as is, as that stored in
the calling thread, and it is easier for an application
to determine if the calling thread needs additional stack
space.

(cherry picked from commit 11c3ee12a89a1c88db47a10fdcb62ec301857c37)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 11c3ee12a89a1c88db47a10fdcb62ec301857c37
Cr-Build-Id: 8716212100230001793
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8716212100230001793
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ibabf044a95f0ea223a6fd3e6f7d9d1533f2a44d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6500087
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>