drivers: disk: nvme: fix warnings
There are warnings generated within NVMe. `prp` is a `void *` which
is 32bits wide on 32bit systems. This adds a cast to first cast it
to a `uintptr_t` and then casts it to a `uint64_t` to supress the
warning.

This also fix an issue where `int n_prp` is defined under a case
statement. This adds the { } around the block underneath it.

(cherry picked from commit 05db1563c53de43ba32ab00c013aba2137a28fed)

Original-Signed-off-by: Ryan McClelland <ryanmcclelland@meta.com>
GitOrigin-RevId: 05db1563c53de43ba32ab00c013aba2137a28fed
Cr-Build-Id: 8716121505362150401
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8716121505362150401
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I7c7abd279f9d8807b3a9c656e4cd58198400564b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6503718
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>