drivers: sensors: lsm6dsv16x: Fix bug in decoder
The lsm6dsv16x driver is using an incorrect value for the index
into the acceleration range. This results in using the wrong
range. The issue was found to be a bug in a MACRO in
lsm6dsv16x_decoder.h that was used to find the index in the
acceleration range. The MACRO was improperly counting zeros from the
leading edge, instead of the trailing edge.

(cherry picked from commit ead1cfb1aaadbf8c137c9af8ea839828c7b2dfa3)

Original-Signed-off-by: Derek Valleroy <derekvalleroy@fb.com>
GitOrigin-RevId: ead1cfb1aaadbf8c137c9af8ea839828c7b2dfa3
Cr-Build-Id: 8717910794353548529
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8717910794353548529
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I28813e5efbc1ebe0ed83507af4787b0dfeecf54a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6450443
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>