Bluetooth: Classic: HFP_AG: Don't change call status if SLC broken
In current implementation, the active/hold call will be terminated
and notify the upper layer the status change of calls when the SLC
is broken. But the calls should not be terminated in this case.

Do not terminate the calls and only clear the status of all calls.
And disconnect the SCO connection if it is established.

(cherry picked from commit 03ca91fef49c0ebd099cdf9b78120b29fcd12c27)

Original-Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>
GitOrigin-RevId: 03ca91fef49c0ebd099cdf9b78120b29fcd12c27
Cr-Build-Id: 8717616354299502449
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8717616354299502449
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I63e0999c09cecb802178f32160318caab021ffc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6454016
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Tested-by: Ting Shen <phoenixshen@chromium.org>