tests: Bluetooth: BSIM: Store last sent opcode for BT tester
The BT Tester will not provide the opcode in the case of a
BTP error. To help log this for developers, we store the
opcode of the currently outstanding command, so that we can
log it.

(cherry picked from commit abad49b0d34672653d9937b333f453d69c237233)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: abad49b0d34672653d9937b333f453d69c237233
Cr-Build-Id: 8715940311004562449
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8715940311004562449
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I658a48341fb70ce114f8e63c13e1917da97bd879
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6509616
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>