boards: Support Teensy Micromod
The Sparkfun Teensy Micromod was designed by Paul of PJRC
and manufactured and sold by Sparkfun.com

The functionality of this board is very similar to the Teensy 4.

It has the same basic pin definitions of the Teensy 4 but added
6 additional IO pins. Which added a lot of possibilities to use
the flexio system in new ways, such as for camera input.

It was unclear if the board should be added here or all of this
duplicated under Sparkfun, but now Sparkfun is also now responsible for
 the manufacturing and sales of the Teensy 4 and Teensy 4.1, here feels
 like a good place.

zephyr directly, as most of my experiences with Zephyr is done through
the current WIP Arduino Wrappers for some of their boards.

So far I duplicated the Teensy 4.1 files, then modified, like removed
the Ethernet and added the I2C2 object.

I updated the flash information as these come with a 128mbit version
versus the 64 on T4.1 or 16 on T4.

So far I have only test blink.
plus logging, plus blinking.

(cherry picked from commit d48f2f49a40e37613ea58ae2154b5f596064c1c7)

Original-Note: I am mainly doing this as a learning experience on how to use
Original-plus: cdc-acm example, plus adding gpio to it,
Original-Signed-off-by: Kurt Eckhardt <kurte@rockisland.com>
GitOrigin-RevId: d48f2f49a40e37613ea58ae2154b5f596064c1c7
Cr-Build-Id: 8715600573159081185
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8715600573159081185
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I8300478ff68c9de783431d6cac1983505d62ec7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6515932
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Tristan Honscheid <honscheid@google.com>