drivers: clock: stm32H5 clock control of the PLL in MemoryMapped mode
Do not disable the PLL clock if it sources the XSPI and if the external
flash is executing in Place. After mcuboot reset, the code is executed
on the external flash, through the xspi.
The CONFIG_STM32_APP_IN_EXT_FLASH is set and will avoid re-config
of the PLL which is sourcing the XSPI peripheral. When eXecuting in Place
on this external NOR, it must not disable its own clock source (PLL).

(cherry picked from commit a63683aa956f31b89e1859750eb5fbbe606c82cf)

Original-Signed-off-by: Francois Ramu <francois.ramu@st.com>
GitOrigin-RevId: a63683aa956f31b89e1859750eb5fbbe606c82cf
Cr-Build-Id: 8708919045324718033
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8708919045324718033
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: If2bb658bf17fdafcdc873e89abff22eb470ab4c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6772615
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>