net: gPTP: Fix pointer type in gptp_add_port
The number of ports in gptp_domain.default_ds.nb_ports is a uint8_t.
A pointer to it is passed to gptp_add_port.
However, in this method, the pointer is cast to an int pointer.
The C compiler generates an int-size store for this.
In addition to potentially overwriting adjacent attributes, on platforms
such as RISC-V that do not support unaligned accesses, this causes an
exception on access.
This commit casts nb_ports to the correct type, uint8_t.

(cherry picked from commit 421a7047e5223630e7cb3b4a13e08aa572f50a39)

Original-Signed-off-by: Eric Ackermann <eric.ackermann@cispa.de>
GitOrigin-RevId: 421a7047e5223630e7cb3b4a13e08aa572f50a39
Cr-Build-Id: 8717503108539528641
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8717503108539528641
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I8598132acdf6c39e83a4e30bdb61b2ab47877a6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6460056
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Tested-by: Ting Shen <phoenixshen@chromium.org>