boards: silabs: siwx91x: Expose real layout of the flash
The Network Coprocessor on SiWx91x owns a large part of the flash. Zephyr
is not expected to access to theses areas.

However, it is still technically possible to access these. In addition, we
prefer the DTS contains a comprehensive and transparent description of the
hardware. So update the DTS with the real partitioning of the SoC.

Reference documentation is available here[1].

[1]: https://www.silabs.com/documents/public/application-notes/
     an1416-siwx917-soc-memory-map.pdf

(cherry picked from commit 53375e95ba150f478fa4f3d37b599964e0cdf50e)

Original-Signed-off-by: Jérôme Pouiller <jerome.pouiller@silabs.com>
GitOrigin-RevId: 53375e95ba150f478fa4f3d37b599964e0cdf50e
Cr-Build-Id: 8708692553088045921
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8708692553088045921
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ide2f75f06ba7d00b6aa8bf55eb3a9cf838fcc5a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6776263
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>